- Moved the send mails methods into utils/mail.py - Added _invalid_tries to session in order to limit the tries of invalid change password - Added INVALID_TRIES variable to the config
