+++ lms/lmsweb/forms/change_password.py
@@ -1,7 +1,10 @@
+from flask import session
 from flask_babel import gettext as _  # type: ignore
 from flask_wtf import FlaskForm
 from wtforms import PasswordField
 from wtforms.validators import EqualTo, InputRequired, Length, ValidationError
+
+from lms.lmsweb.config import INVALID_TRIES
 
 
 class ChangePasswordForm(FlaskForm):
@@ -23,5 +26,8 @@ class ChangePasswordForm(FlaskForm):
         self.user = user
 
     def validate_current_password(self, field):
+        if session['_invalid_tries'] == INVALID_TRIES:
+            raise ValidationError(_('הזנת סיסמה שגוייה מספר רב מידי של פעמים'))
         if not self.user.is_password_valid(field.data):
+            session['_invalid_tries'] += 1
             raise ValidationError(_('הסיסמה הנוכחית שהוזנה שגויה'))
