+++ lms/templates/user.html
@@ -15,7 +15,9 @@
       <div class="user-actions {{ direction }}">
         <h2>{{ _('פעולות:') }}</h2>
         <div id="change-password-user">
-          <a href="/change-password" class="btn btn-primary btn-sm" role="button">{{ _('שנה סיסמה') }}</a>
+          <ul>
+            <li><a href="/change-password" role="button">{{ _('שנה סיסמה') }}</a></li>
+          </ul>
         </div>
       </div>
       <div class="user-exercises {{ direction }}">
